new implementations for the game class
