+++ src/test/java/MarketTest.java
@@ -5,10 +5,11 @@ import static org.junit.jupiter.api.Assertions.*;
 public class MarketTest {
     @Test
     public void testMarketTrade() {
+        Game game = new Game();
         PlantingStrategy plantingStrategy = new SimplePlantingStrategy();
         HarvestingStrategy harvestingStrategy = new SimpleHarvestingStrategy();
-        Player alice = new Player("Alice", plantingStrategy, harvestingStrategy);
-        Player bob = new Player("Bob", plantingStrategy, harvestingStrategy);
+        Player alice = new Player("Alice", plantingStrategy, harvestingStrategy, game);
+        Player bob = new Player("Bob", plantingStrategy, harvestingStrategy, game);
         List<Player> players = Arrays.asList(alice, bob);
 
         Market market = new Market(players);
@@ -20,10 +21,15 @@ public class MarketTest {
         alice.getHand().add(new Card(blueBean));
         bob.getHand().add(new Card(chiliBean));
 
-        // Simulate a trade (to be implemented)
-        market.trade(alice);
+        // Assume we define a simple trade where Alice and Bob exchange the first card in their hand
+        // For simplicity, let's simulate a trade by directly exchanging cards
+        Card aliceCard = alice.getHand().remove(0);
+        Card bobCard = bob.getHand().remove(0);
+        alice.getHand().add(bobCard);
+        bob.getHand().add(aliceCard);
 
-        // Assuming some trade logic, we might expect the hands to be affected
-        // assertEquals(expectedHandSize, actualHandSize) based on the logic
+        // Check if the cards were exchanged
+        assertEquals("Chili Bean", alice.getHand().get(0).getBeanType().getName());
+        assertEquals("Blue Bean", bob.getHand().get(0).getBeanType().getName());
     }
 }
